Start restic rest-server service units on boot
With `StopWhenUnneeded=yes` the restic rest-server service units failed
to start after the deployment of this role:

  Sep 22 17:13:55 backup01 systemd[1]: Started Restic REST Server (/srv/backups/restic/archive).
  Sep 22 17:13:55 backup01 systemd[1]: rest-server-2b6d7d.service: Unit not needed anymore. Stopping.
  Sep 22 17:13:55 backup01 systemd[1]: Stopping Restic REST Server (/srv/backups/restic/archive)...
  Sep 22 17:13:55 backup01 systemd[1]: Stopped Restic REST Server (/srv/backups/restic/archive).

"Unit not needed anymore" is caused by StopWhenUnneeded=yes`. Quoting systemd.unit(5):

  StopWhenUnneeded=

      Takes a boolean argument. If true, this unit will be stopped when
      it is no longer used. Note that, in order to minimize the work to
      be executed, systemd will not stop units by default unless they
      are conflicting with other units, or the user explicitly requested
      their shut down. If this option is set, a unit will be
      automatically cleaned up if no other active unit requires it.
      Defaults to false.

The rest-server.target can be used to start all servers at once,
but enabling and starting the rest-server service units directly seem to be better.

To reflect that change, changed the name of the variable
`restic_rest_target_enable` to `restic_rest_service_enable` which is still set
to false by default.

So basically we do not start the rest-server.target via ansible anymore,
instead we enable and start the rest-server service units (if
restic_rest_service_enable is true).